Restructure the project - Created folder for controllers - Created folder for services
