+++ app/index.html
@@ -58,7 +58,7 @@
 				<p class="lead">
 					Resumo do balanço financeiro(impostos e destino dos gastos) do seu município baseado em
 					dados do <a href="http://transparencia.tce.sp.gov.br/">TCE de
-						São Paulo</a>. 
+						São Paulo</a>.
 				</p>
                 </div>
 			</div>
@@ -89,7 +89,7 @@
             </div>
             <!-- END REVENUE -->
             <!-- START EXPENSES -->
-            <div class="row well expensesDiv"">
+            <div class="row well expensesDiv">
               <h3 class="text-center sectionHeader">Despesas</h3>
               <div id="expensesChart" class="col-lg-12"></div>
             </div>
@@ -164,7 +164,9 @@
 	<script src="./bower_components/jquery/dist/jquery.min.js"></script>
 	<script src="./bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
     <script src="./bower_components/highcharts/highcharts.js"></script>
-	<script src="./js/controller.js"></script>
+	<script src="./js/app.js"></script>
+    <script src="./js/services/expense.service.js"></script>
+    <script src="./js/controllers/balanco.controller.js"></script>
 	<!-- endbuild -->
 </body>
 </html>
