+++ app/js/app.js
@@ -1,0 +1,3 @@
+(function() {
+    angular.module('BalancoApp', []);
+}());
